Fix needless GC in TerrainUtility
Updated TileCoord in TerrainUtility to not use default struct comparison. The default struct comparison creates some allocation whenever it is used for hashing purposes, as in a dictionary. TerrainUtility uses the TileCoord struct extensively to lookup terrain tiles. In a test scenario where 1000 TileCoords are added and looked for in a dictionary, this creates some 23MB of garbage. This changes should reduce that to zero.